perf: Change from "Identify Headers" to "Styled Components" to improve load page speed
It seems what i as using, which was a function that recognizes mobile devices thought its header, was slowing down the page load. It was used to show images that was mobile friendly.
So i changed to Css in Js solution, like Styled Components. It gave the same results, but with quickier load speed.